requirements.txt
beautifulsoup4==4.12.2
clearbit==0.1.7
fake_useragent==1.1.3
fastapi==0.95.2
fuzzywuzzy==0.18.0
googletrans==4.0.0rc1
Jinja2==3.1.2
linkedin_api==2.0.1
openai==0.25.0
pycountry==20.7.3
pydantic==1.8.2
python-dotenv==1.0.0
python-Levenshtein==0.20.9
python-multipart==0.0.6
requests==2.28.2
tldextract==3.4.2
uvicorn==0.16.0

